+++ connector-api/src/main/java/it/pagopa/selfcare/mscore/constant/DelegationState.java
@@ -1,0 +1,6 @@
+package it.pagopa.selfcare.mscore.constant;
+
+public enum DelegationState {
+    ACTIVE,
+    DELETED,
+}
