+++ connector-api/src/main/java/it/pagopa/selfcare/mscore/model/institution/Institution.java
@@ -48,5 +48,6 @@ public class Institution {
     private String rootParentId;
     private String parentDescription;
     private PaAttributes paAttributes;
+    private boolean delegation;
 
 }
