+++ connector/dao/src/main/java/it/pagopa/selfcare/mscore/connector/dao/model/DelegationEntity.java
@@ -1,11 +1,14 @@
 package it.pagopa.selfcare.mscore.connector.dao.model;
 
+import it.pagopa.selfcare.mscore.constant.DelegationState;
 import it.pagopa.selfcare.mscore.constant.DelegationType;
 import lombok.Data;
 import lombok.experimental.FieldNameConstants;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.Sharded;
+
+import java.time.OffsetDateTime;
 
 @Data
 @Document("Delegations")
@@ -22,5 +25,8 @@ public class DelegationEntity {
     private String to;
     private String productId;
     private DelegationType type;
+    private DelegationState status;
+    private OffsetDateTime createdAt;
+    private OffsetDateTime updatedAt;
 
 }
