+++ connector/dao/src/test/java/it/pagopa/selfcare/mscore/connector/dao/TokenConnectorImplTest.java
@@ -60,475 +60,10 @@ class TokenConnectorImplTest {
     @Captor
     ArgumentCaptor<Pageable> pageableArgumentCaptor;
 
-    /**
-     * Method under test: {@link TokenConnectorImpl#findAll()}
-     */
-    @Test
-    void testFindAll() {
-        when(tokenRepository.findAll()).thenReturn(new ArrayList<>());
-        assertTrue(tokenConnectorImpl.findAll().isEmpty());
-        verify(tokenRepository).findAll();
-    }
-
-    /**
-     * Method under test: {@link TokenConnectorImpl#findAll()}
-     */
-    @Test
-    void testFindAll2() {
-        DataProtectionOfficerEntity dataProtectionOfficerEntity = new DataProtectionOfficerEntity();
-        dataProtectionOfficerEntity.setAddress("42 Main St");
-        dataProtectionOfficerEntity.setEmail("jane.doe@example.org");
-        dataProtectionOfficerEntity.setPec("Pec");
-
-        PaymentServiceProviderEntity paymentServiceProviderEntity = new PaymentServiceProviderEntity();
-        paymentServiceProviderEntity.setAbiCode("Abi Code");
-        paymentServiceProviderEntity.setBusinessRegisterNumber("42");
-        paymentServiceProviderEntity.setLegalRegisterName("Legal Register Name");
-        paymentServiceProviderEntity.setLegalRegisterNumber("42");
-        paymentServiceProviderEntity.setVatNumberGroup(true);
-
-        InstitutionUpdateEntity institutionUpdateEntity = new InstitutionUpdateEntity();
-        institutionUpdateEntity.setAddress("42 Main St");
-        institutionUpdateEntity.setBusinessRegisterPlace("Business Register Place");
-        institutionUpdateEntity.setDataProtectionOfficer(dataProtectionOfficerEntity);
-        institutionUpdateEntity.setDescription("The characteristics of someone or something");
-        institutionUpdateEntity.setDigitalAddress("42 Main St");
-        institutionUpdateEntity.setGeographicTaxonomies(new ArrayList<>());
-        institutionUpdateEntity.setImported(true);
-        institutionUpdateEntity.setInstitutionType(InstitutionType.PA);
-        institutionUpdateEntity.setPaymentServiceProvider(paymentServiceProviderEntity);
-        institutionUpdateEntity.setRea("Rea");
-        institutionUpdateEntity.setShareCapital("Share Capital");
-        institutionUpdateEntity.setSupportEmail("jane.doe@example.org");
-        institutionUpdateEntity.setSupportPhone("6625550144");
-        institutionUpdateEntity.setTaxCode("Tax Code");
-        institutionUpdateEntity.setZipCode("21654");
-
-        TokenEntity tokenEntity = new TokenEntity();
-        tokenEntity.setChecksum("Checksum");
-        tokenEntity.setDeletedAt(null);
-        tokenEntity.setContractSigned("Contract Signed");
-        tokenEntity.setContractTemplate("Contract Template");
-        tokenEntity.setCreatedAt(null);
-        tokenEntity.setExpiringDate(null);
-        tokenEntity.setId("42");
-        tokenEntity.setInstitutionId("42");
-        tokenEntity.setInstitutionUpdate(institutionUpdateEntity);
-        tokenEntity.setProductId("42");
-        tokenEntity.setStatus(RelationshipState.PENDING);
-        tokenEntity.setType(TokenType.INSTITUTION);
-        tokenEntity.setUpdatedAt(null);
-        tokenEntity.setUsers(new ArrayList<>());
-
-        ArrayList<TokenEntity> tokenEntityList = new ArrayList<>();
-        tokenEntityList.add(tokenEntity);
-        when(tokenRepository.findAll()).thenReturn(tokenEntityList);
-        assertEquals(1, tokenConnectorImpl.findAll().size());
-        verify(tokenRepository).findAll();
-    }
-
-    /**
-     * Method under test: {@link TokenConnectorImpl#findAll()}
-     */
-    @Test
-    void testFindAll3() {
-        DataProtectionOfficerEntity dataProtectionOfficerEntity = new DataProtectionOfficerEntity();
-        dataProtectionOfficerEntity.setAddress("42 Main St");
-        dataProtectionOfficerEntity.setEmail("jane.doe@example.org");
-        dataProtectionOfficerEntity.setPec("Pec");
-
-        PaymentServiceProviderEntity paymentServiceProviderEntity = new PaymentServiceProviderEntity();
-        paymentServiceProviderEntity.setAbiCode("Abi Code");
-        paymentServiceProviderEntity.setBusinessRegisterNumber("42");
-        paymentServiceProviderEntity.setLegalRegisterName("Legal Register Name");
-        paymentServiceProviderEntity.setLegalRegisterNumber("42");
-        paymentServiceProviderEntity.setVatNumberGroup(true);
-
-        InstitutionUpdateEntity institutionUpdateEntity = new InstitutionUpdateEntity();
-        institutionUpdateEntity.setAddress("42 Main St");
-        institutionUpdateEntity.setBusinessRegisterPlace("Business Register Place");
-        institutionUpdateEntity.setDataProtectionOfficer(dataProtectionOfficerEntity);
-        institutionUpdateEntity.setDescription("The characteristics of someone or something");
-        institutionUpdateEntity.setDigitalAddress("42 Main St");
-        institutionUpdateEntity.setGeographicTaxonomies(new ArrayList<>());
-        institutionUpdateEntity.setImported(true);
-        institutionUpdateEntity.setInstitutionType(InstitutionType.PA);
-        institutionUpdateEntity.setPaymentServiceProvider(paymentServiceProviderEntity);
-        institutionUpdateEntity.setRea("Rea");
-        institutionUpdateEntity.setShareCapital("Share Capital");
-        institutionUpdateEntity.setSupportEmail("jane.doe@example.org");
-        institutionUpdateEntity.setSupportPhone("6625550144");
-        institutionUpdateEntity.setTaxCode("Tax Code");
-        institutionUpdateEntity.setZipCode("21654");
-
-        TokenEntity tokenEntity = new TokenEntity();
-        tokenEntity.setChecksum("Checksum");
-        tokenEntity.setDeletedAt(null);
-        tokenEntity.setContractSigned("Contract Signed");
-        tokenEntity.setContractTemplate("Contract Template");
-        tokenEntity.setCreatedAt(null);
-        tokenEntity.setExpiringDate(null);
-        tokenEntity.setId("42");
-        tokenEntity.setInstitutionId("42");
-        tokenEntity.setInstitutionUpdate(institutionUpdateEntity);
-        tokenEntity.setProductId("42");
-        tokenEntity.setStatus(RelationshipState.PENDING);
-        tokenEntity.setType(TokenType.INSTITUTION);
-        tokenEntity.setUpdatedAt(null);
-        tokenEntity.setUsers(new ArrayList<>());
-
-        DataProtectionOfficerEntity dataProtectionOfficerEntity1 = new DataProtectionOfficerEntity();
-        dataProtectionOfficerEntity1.setAddress("17 High St");
-        dataProtectionOfficerEntity1.setEmail("john.smith@example.org");
-        dataProtectionOfficerEntity1
-                .setPec("it.pagopa.selfcare.mscore.connector.dao.model.inner.DataProtectionOfficerEntity");
-
-        PaymentServiceProviderEntity paymentServiceProviderEntity1 = new PaymentServiceProviderEntity();
-        paymentServiceProviderEntity1
-                .setAbiCode("it.pagopa.selfcare.mscore.connector.dao.model.inner.PaymentServiceProviderEntity");
-        paymentServiceProviderEntity1.setBusinessRegisterNumber("Business Register Number");
-        paymentServiceProviderEntity1
-                .setLegalRegisterName("it.pagopa.selfcare.mscore.connector.dao.model.inner.PaymentServiceProviderEntity");
-        paymentServiceProviderEntity1.setLegalRegisterNumber("Legal Register Number");
-        paymentServiceProviderEntity1.setVatNumberGroup(false);
-
-        InstitutionUpdateEntity institutionUpdateEntity1 = new InstitutionUpdateEntity();
-        institutionUpdateEntity1.setAddress("17 High St");
-        institutionUpdateEntity1
-                .setBusinessRegisterPlace("it.pagopa.selfcare.mscore.connector.dao.model.inner.InstitutionUpdateEntity");
-        institutionUpdateEntity1.setDataProtectionOfficer(dataProtectionOfficerEntity1);
-        institutionUpdateEntity1.setDescription("Description");
-        institutionUpdateEntity1.setDigitalAddress("17 High St");
-        institutionUpdateEntity1.setGeographicTaxonomies(new ArrayList<>());
-        institutionUpdateEntity1.setImported(false);
-        institutionUpdateEntity1.setInstitutionType(InstitutionType.PG);
-        institutionUpdateEntity1.setPaymentServiceProvider(paymentServiceProviderEntity1);
-        institutionUpdateEntity1.setRea("it.pagopa.selfcare.mscore.connector.dao.model.inner.InstitutionUpdateEntity");
-        institutionUpdateEntity1
-                .setShareCapital("it.pagopa.selfcare.mscore.connector.dao.model.inner.InstitutionUpdateEntity");
-        institutionUpdateEntity1.setSupportEmail("john.smith@example.org");
-        institutionUpdateEntity1.setSupportPhone("8605550118");
-        institutionUpdateEntity1
-                .setTaxCode("it.pagopa.selfcare.mscore.connector.dao.model.inner.InstitutionUpdateEntity");
-        institutionUpdateEntity1.setZipCode("OX1 1PT");
-
-        TokenEntity tokenEntity1 = new TokenEntity();
-        tokenEntity1.setChecksum("it.pagopa.selfcare.mscore.connector.dao.model.TokenEntity");
-        tokenEntity1.setDeletedAt(null);
-        tokenEntity1.setContractSigned("it.pagopa.selfcare.mscore.connector.dao.model.TokenEntity");
-        tokenEntity1.setContractTemplate("it.pagopa.selfcare.mscore.connector.dao.model.TokenEntity");
-        tokenEntity1.setCreatedAt(null);
-        tokenEntity1.setExpiringDate(null);
-        tokenEntity1.setId("Id");
-        tokenEntity1.setInstitutionId("Institution Id");
-        tokenEntity1.setInstitutionUpdate(institutionUpdateEntity1);
-        tokenEntity1.setProductId("Product Id");
-        tokenEntity1.setStatus(RelationshipState.ACTIVE);
-        tokenEntity1.setType(TokenType.LEGALS);
-        tokenEntity1.setUpdatedAt(null);
-        tokenEntity1.setUsers(new ArrayList<>());
-
-        ArrayList<TokenEntity> tokenEntityList = new ArrayList<>();
-        tokenEntityList.add(tokenEntity1);
-        tokenEntityList.add(tokenEntity);
-        when(tokenRepository.findAll()).thenReturn(tokenEntityList);
-        assertEquals(2, tokenConnectorImpl.findAll().size());
-        verify(tokenRepository).findAll();
-    }
-
-    /**
-     * Method under test: {@link TokenConnectorImpl#findAll()}
-     */
-    @Test
-    void testFindAll4() {
-        when(tokenRepository.findAll()).thenThrow(new ResourceNotFoundException("An error occurred", "Code"));
-        assertThrows(ResourceNotFoundException.class, () -> tokenConnectorImpl.findAll());
-        verify(tokenRepository).findAll();
-    }
-
-    /**
-     * Method under test: {@link TokenConnectorImpl#findAll()}
-     */
-    @Test
-    void testFindAll5() {
-        DataProtectionOfficerEntity dataProtectionOfficerEntity = new DataProtectionOfficerEntity();
-        dataProtectionOfficerEntity.setAddress("42 Main St");
-        dataProtectionOfficerEntity.setEmail("jane.doe@example.org");
-        dataProtectionOfficerEntity.setPec("Pec");
-
-        PaymentServiceProviderEntity paymentServiceProviderEntity = new PaymentServiceProviderEntity();
-        paymentServiceProviderEntity.setAbiCode("Abi Code");
-        paymentServiceProviderEntity.setBusinessRegisterNumber("42");
-        paymentServiceProviderEntity.setLegalRegisterName("Legal Register Name");
-        paymentServiceProviderEntity.setLegalRegisterNumber("42");
-        paymentServiceProviderEntity.setVatNumberGroup(true);
-
-        InstitutionUpdateEntity institutionUpdateEntity = new InstitutionUpdateEntity();
-        institutionUpdateEntity.setAddress("42 Main St");
-        institutionUpdateEntity.setBusinessRegisterPlace("Business Register Place");
-        institutionUpdateEntity.setDataProtectionOfficer(dataProtectionOfficerEntity);
-        institutionUpdateEntity.setDescription("The characteristics of someone or something");
-        institutionUpdateEntity.setDigitalAddress("42 Main St");
-        institutionUpdateEntity.setGeographicTaxonomies(new ArrayList<>());
-        institutionUpdateEntity.setImported(true);
-        institutionUpdateEntity.setInstitutionType(InstitutionType.PA);
-        institutionUpdateEntity.setPaymentServiceProvider(paymentServiceProviderEntity);
-        institutionUpdateEntity.setRea("Rea");
-        institutionUpdateEntity.setShareCapital("Share Capital");
-        institutionUpdateEntity.setSupportEmail("jane.doe@example.org");
-        institutionUpdateEntity.setSupportPhone("6625550144");
-        institutionUpdateEntity.setTaxCode("Tax Code");
-        institutionUpdateEntity.setZipCode("21654");
-
-        DataProtectionOfficerEntity dataProtectionOfficerEntity1 = new DataProtectionOfficerEntity();
-        dataProtectionOfficerEntity1.setAddress("42 Main St");
-        dataProtectionOfficerEntity1.setEmail("jane.doe@example.org");
-        dataProtectionOfficerEntity1.setPec("Pec");
-
-        PaymentServiceProviderEntity paymentServiceProviderEntity1 = new PaymentServiceProviderEntity();
-        paymentServiceProviderEntity1.setAbiCode("Abi Code");
-        paymentServiceProviderEntity1.setBusinessRegisterNumber("42");
-        paymentServiceProviderEntity1.setLegalRegisterName("Legal Register Name");
-        paymentServiceProviderEntity1.setLegalRegisterNumber("42");
-        paymentServiceProviderEntity1.setVatNumberGroup(true);
-
-        DataProtectionOfficerEntity dataProtectionOfficerEntity2 = new DataProtectionOfficerEntity();
-        dataProtectionOfficerEntity2.setAddress("42 Main St");
-        dataProtectionOfficerEntity2.setEmail("jane.doe@example.org");
-        dataProtectionOfficerEntity2.setPec("Pec");
-
-        PaymentServiceProviderEntity paymentServiceProviderEntity2 = new PaymentServiceProviderEntity();
-        paymentServiceProviderEntity2.setAbiCode("Abi Code");
-        paymentServiceProviderEntity2.setBusinessRegisterNumber("42");
-        paymentServiceProviderEntity2.setLegalRegisterName("Legal Register Name");
-        paymentServiceProviderEntity2.setLegalRegisterNumber("42");
-        paymentServiceProviderEntity2.setVatNumberGroup(true);
-
-        GeoTaxonomyEntity geoTaxonomyEntity = new GeoTaxonomyEntity();
-        geoTaxonomyEntity.setCode("Code");
-        geoTaxonomyEntity.setDesc("The characteristics of someone or something");
-
-        ArrayList<GeoTaxonomyEntity> geoTaxonomyEntityList = new ArrayList<>();
-        geoTaxonomyEntityList.add(geoTaxonomyEntity);
-        InstitutionUpdateEntity institutionUpdateEntity1 = mock(InstitutionUpdateEntity.class);
-        when(institutionUpdateEntity1.isImported()).thenReturn(true);
-        when(institutionUpdateEntity1.getDataProtectionOfficer()).thenReturn(dataProtectionOfficerEntity2);
-        when(institutionUpdateEntity1.getPaymentServiceProvider()).thenReturn(paymentServiceProviderEntity2);
-        when(institutionUpdateEntity1.getAddress()).thenReturn("42 Main St");
-        when(institutionUpdateEntity1.getBusinessRegisterPlace()).thenReturn("Business Register Place");
-        when(institutionUpdateEntity1.getDescription()).thenReturn("The characteristics of someone or something");
-        when(institutionUpdateEntity1.getDigitalAddress()).thenReturn("42 Main St");
-        when(institutionUpdateEntity1.getRea()).thenReturn("Rea");
-        when(institutionUpdateEntity1.getShareCapital()).thenReturn("Share Capital");
-        when(institutionUpdateEntity1.getSupportEmail()).thenReturn("jane.doe@example.org");
-        when(institutionUpdateEntity1.getSupportPhone()).thenReturn("6625550144");
-        when(institutionUpdateEntity1.getTaxCode()).thenReturn("Tax Code");
-        when(institutionUpdateEntity1.getZipCode()).thenReturn("21654");
-        when(institutionUpdateEntity1.getGeographicTaxonomies()).thenReturn(geoTaxonomyEntityList);
-        institutionUpdateEntity1.setAddress("42 Main St");
-        institutionUpdateEntity1.setBusinessRegisterPlace("Business Register Place");
-        institutionUpdateEntity1.setDataProtectionOfficer(dataProtectionOfficerEntity1);
-        institutionUpdateEntity1.setDescription("The characteristics of someone or something");
-        institutionUpdateEntity1.setDigitalAddress("42 Main St");
-        institutionUpdateEntity1.setGeographicTaxonomies(new ArrayList<>());
-        institutionUpdateEntity1.setImported(true);
-        institutionUpdateEntity1.setInstitutionType(InstitutionType.PA);
-        institutionUpdateEntity1.setPaymentServiceProvider(paymentServiceProviderEntity1);
-        institutionUpdateEntity1.setRea("Rea");
-        institutionUpdateEntity1.setShareCapital("Share Capital");
-        institutionUpdateEntity1.setSupportEmail("jane.doe@example.org");
-        institutionUpdateEntity1.setSupportPhone("6625550144");
-        institutionUpdateEntity1.setTaxCode("Tax Code");
-        institutionUpdateEntity1.setZipCode("21654");
-        TokenEntity tokenEntity = mock(TokenEntity.class);
-        when(tokenEntity.getInstitutionUpdate()).thenReturn(institutionUpdateEntity1);
-        when(tokenEntity.getStatus()).thenReturn(RelationshipState.PENDING);
-        when(tokenEntity.getChecksum()).thenReturn("Checksum");
-        when(tokenEntity.getContractSigned()).thenReturn("Contract Signed");
-        when(tokenEntity.getContractTemplate()).thenReturn("Contract Template");
-        when(tokenEntity.getId()).thenReturn("42");
-        when(tokenEntity.getInstitutionId()).thenReturn("42");
-        when(tokenEntity.getProductId()).thenReturn("42");
-        when(tokenEntity.getCreatedAt()).thenReturn(null);
-        when(tokenEntity.getExpiringDate()).thenReturn(null);
-        when(tokenEntity.getUpdatedAt()).thenReturn(null);
-        when(tokenEntity.getUsers()).thenReturn(new ArrayList<>());
-        tokenEntity.setChecksum("Checksum");
-        tokenEntity.setDeletedAt(null);
-        tokenEntity.setContractSigned("Contract Signed");
-        tokenEntity.setContractTemplate("Contract Template");
-        tokenEntity.setCreatedAt(null);
-        tokenEntity.setExpiringDate(null);
-        tokenEntity.setId("42");
-        tokenEntity.setInstitutionId("42");
-        tokenEntity.setInstitutionUpdate(institutionUpdateEntity);
-        tokenEntity.setProductId("42");
-        tokenEntity.setStatus(RelationshipState.PENDING);
-        tokenEntity.setType(TokenType.INSTITUTION);
-        tokenEntity.setUpdatedAt(null);
-        tokenEntity.setUsers(new ArrayList<>());
-
-        ArrayList<TokenEntity> tokenEntityList = new ArrayList<>();
-        tokenEntityList.add(tokenEntity);
-        when(tokenRepository.findAll()).thenReturn(tokenEntityList);
-        assertEquals(1, tokenConnectorImpl.findAll().size());
-    }
-
-    /**
-     * Method under test: {@link TokenConnectorImpl#findAll()}
-     */
-    @Test
-    void testFindAll6() {
-        DataProtectionOfficerEntity dataProtectionOfficerEntity = new DataProtectionOfficerEntity();
-        dataProtectionOfficerEntity.setAddress("42 Main St");
-        dataProtectionOfficerEntity.setEmail("jane.doe@example.org");
-        dataProtectionOfficerEntity.setPec("Pec");
-
-        PaymentServiceProviderEntity paymentServiceProviderEntity = new PaymentServiceProviderEntity();
-        paymentServiceProviderEntity.setAbiCode("Abi Code");
-        paymentServiceProviderEntity.setBusinessRegisterNumber("42");
-        paymentServiceProviderEntity.setLegalRegisterName("Legal Register Name");
-        paymentServiceProviderEntity.setLegalRegisterNumber("42");
-        paymentServiceProviderEntity.setVatNumberGroup(true);
-
-        InstitutionUpdateEntity institutionUpdateEntity = new InstitutionUpdateEntity();
-        institutionUpdateEntity.setAddress("42 Main St");
-        institutionUpdateEntity.setBusinessRegisterPlace("Business Register Place");
-        institutionUpdateEntity.setDataProtectionOfficer(dataProtectionOfficerEntity);
-        institutionUpdateEntity.setDescription("The characteristics of someone or something");
-        institutionUpdateEntity.setDigitalAddress("42 Main St");
-        institutionUpdateEntity.setGeographicTaxonomies(new ArrayList<>());
-        institutionUpdateEntity.setImported(true);
-        institutionUpdateEntity.setInstitutionType(InstitutionType.PA);
-        institutionUpdateEntity.setPaymentServiceProvider(paymentServiceProviderEntity);
-        institutionUpdateEntity.setRea("Rea");
-        institutionUpdateEntity.setShareCapital("Share Capital");
-        institutionUpdateEntity.setSupportEmail("jane.doe@example.org");
-        institutionUpdateEntity.setSupportPhone("6625550144");
-        institutionUpdateEntity.setTaxCode("Tax Code");
-        institutionUpdateEntity.setZipCode("21654");
-
-        DataProtectionOfficerEntity dataProtectionOfficerEntity1 = new DataProtectionOfficerEntity();
-        dataProtectionOfficerEntity1.setAddress("42 Main St");
-        dataProtectionOfficerEntity1.setEmail("jane.doe@example.org");
-        dataProtectionOfficerEntity1.setPec("Pec");
-
-        PaymentServiceProviderEntity paymentServiceProviderEntity1 = new PaymentServiceProviderEntity();
-        paymentServiceProviderEntity1.setAbiCode("Abi Code");
-        paymentServiceProviderEntity1.setBusinessRegisterNumber("42");
-        paymentServiceProviderEntity1.setLegalRegisterName("Legal Register Name");
-        paymentServiceProviderEntity1.setLegalRegisterNumber("42");
-        paymentServiceProviderEntity1.setVatNumberGroup(true);
-
-        DataProtectionOfficerEntity dataProtectionOfficerEntity2 = new DataProtectionOfficerEntity();
-        dataProtectionOfficerEntity2.setAddress("42 Main St");
-        dataProtectionOfficerEntity2.setEmail("jane.doe@example.org");
-        dataProtectionOfficerEntity2.setPec("Pec");
-
-        PaymentServiceProviderEntity paymentServiceProviderEntity2 = new PaymentServiceProviderEntity();
-        paymentServiceProviderEntity2.setAbiCode("Abi Code");
-        paymentServiceProviderEntity2.setBusinessRegisterNumber("42");
-        paymentServiceProviderEntity2.setLegalRegisterName("Legal Register Name");
-        paymentServiceProviderEntity2.setLegalRegisterNumber("42");
-        paymentServiceProviderEntity2.setVatNumberGroup(true);
-
-        GeoTaxonomyEntity geoTaxonomyEntity = new GeoTaxonomyEntity();
-        geoTaxonomyEntity.setCode("Code");
-        geoTaxonomyEntity.setDesc("The characteristics of someone or something");
-
-        GeoTaxonomyEntity geoTaxonomyEntity1 = new GeoTaxonomyEntity();
-        geoTaxonomyEntity1.setCode("it.pagopa.selfcare.mscore.connector.dao.model.inner.GeoTaxonomyEntity");
-        geoTaxonomyEntity1.setDesc("Desc");
-
-        ArrayList<GeoTaxonomyEntity> geoTaxonomyEntityList = new ArrayList<>();
-        geoTaxonomyEntityList.add(geoTaxonomyEntity1);
-        geoTaxonomyEntityList.add(geoTaxonomyEntity);
-        InstitutionUpdateEntity institutionUpdateEntity1 = mock(InstitutionUpdateEntity.class);
-        when(institutionUpdateEntity1.isImported()).thenReturn(true);
-        when(institutionUpdateEntity1.getDataProtectionOfficer()).thenReturn(dataProtectionOfficerEntity2);
-        when(institutionUpdateEntity1.getPaymentServiceProvider()).thenReturn(paymentServiceProviderEntity2);
-        when(institutionUpdateEntity1.getAddress()).thenReturn("42 Main St");
-        when(institutionUpdateEntity1.getBusinessRegisterPlace()).thenReturn("Business Register Place");
-        when(institutionUpdateEntity1.getDescription()).thenReturn("The characteristics of someone or something");
-        when(institutionUpdateEntity1.getDigitalAddress()).thenReturn("42 Main St");
-        when(institutionUpdateEntity1.getRea()).thenReturn("Rea");
-        when(institutionUpdateEntity1.getShareCapital()).thenReturn("Share Capital");
-        when(institutionUpdateEntity1.getSupportEmail()).thenReturn("jane.doe@example.org");
-        when(institutionUpdateEntity1.getSupportPhone()).thenReturn("6625550144");
-        when(institutionUpdateEntity1.getTaxCode()).thenReturn("Tax Code");
-        when(institutionUpdateEntity1.getZipCode()).thenReturn("21654");
-        when(institutionUpdateEntity1.getGeographicTaxonomies()).thenReturn(geoTaxonomyEntityList);
-        institutionUpdateEntity1.setAddress("42 Main St");
-        institutionUpdateEntity1.setBusinessRegisterPlace("Business Register Place");
-        institutionUpdateEntity1.setDataProtectionOfficer(dataProtectionOfficerEntity1);
-        institutionUpdateEntity1.setDescription("The characteristics of someone or something");
-        institutionUpdateEntity1.setDigitalAddress("42 Main St");
-        institutionUpdateEntity1.setGeographicTaxonomies(new ArrayList<>());
-        institutionUpdateEntity1.setImported(true);
-        institutionUpdateEntity1.setInstitutionType(InstitutionType.PA);
-        institutionUpdateEntity1.setPaymentServiceProvider(paymentServiceProviderEntity1);
-        institutionUpdateEntity1.setRea("Rea");
-        institutionUpdateEntity1.setShareCapital("Share Capital");
-        institutionUpdateEntity1.setSupportEmail("jane.doe@example.org");
-        institutionUpdateEntity1.setSupportPhone("6625550144");
-        institutionUpdateEntity1.setTaxCode("Tax Code");
-        institutionUpdateEntity1.setZipCode("21654");
-        TokenEntity tokenEntity = mock(TokenEntity.class);
-        when(tokenEntity.getInstitutionUpdate()).thenReturn(institutionUpdateEntity1);
-        when(tokenEntity.getStatus()).thenReturn(RelationshipState.PENDING);
-        when(tokenEntity.getChecksum()).thenReturn("Checksum");
-        when(tokenEntity.getContractSigned()).thenReturn("Contract Signed");
-        when(tokenEntity.getContractTemplate()).thenReturn("Contract Template");
-        when(tokenEntity.getId()).thenReturn("42");
-        when(tokenEntity.getInstitutionId()).thenReturn("42");
-        when(tokenEntity.getProductId()).thenReturn("42");
-        when(tokenEntity.getCreatedAt()).thenReturn(null);
-        when(tokenEntity.getExpiringDate()).thenReturn(null);
-        when(tokenEntity.getUpdatedAt()).thenReturn(null);
-        when(tokenEntity.getUsers()).thenReturn(new ArrayList<>());
-        tokenEntity.setChecksum("Checksum");
-        tokenEntity.setDeletedAt(null);
-        tokenEntity.setContractSigned("Contract Signed");
-        tokenEntity.setContractTemplate("Contract Template");
-        tokenEntity.setCreatedAt(null);
-        tokenEntity.setExpiringDate(null);
-        tokenEntity.setId("42");
-        tokenEntity.setInstitutionId("42");
-        tokenEntity.setInstitutionUpdate(institutionUpdateEntity);
-        tokenEntity.setProductId("42");
-        tokenEntity.setStatus(RelationshipState.PENDING);
-        tokenEntity.setType(TokenType.INSTITUTION);
-        tokenEntity.setUpdatedAt(null);
-        tokenEntity.setUsers(new ArrayList<>());
-
-        ArrayList<TokenEntity> tokenEntityList = new ArrayList<>();
-        tokenEntityList.add(tokenEntity);
-        when(tokenRepository.findAll()).thenReturn(tokenEntityList);
-        assertEquals(1, tokenConnectorImpl.findAll().size());
-        verify(tokenRepository).findAll();
-    }
-
     @Test
     void deleteById() {
         doNothing().when(tokenRepository).deleteById(any());
         Assertions.assertDoesNotThrow(() -> tokenConnectorImpl.deleteById("507f1f77bcf86cd799439011"));
-    }
-
-    @Test
-    void findActiveContractTest() {
-        TokenEntity token = new TokenEntity();
-        token.setId("507f1f77bcf86cd799439011");
-        token.setUsers(new ArrayList<>());
-        when(tokenRepository.find(any(), any())).thenReturn(List.of(token));
-        Token tokenResp = tokenConnectorImpl.findActiveContract("42", "42", "42");
-        Assertions.assertEquals("507f1f77bcf86cd799439011", tokenResp.getId());
-    }
-
-    @Test
-    void findActiveContractTest1() {
-        when(tokenRepository.find(any(), any())).thenReturn(new ArrayList<>());
-        assertThrows(ResourceNotFoundException.class, () -> tokenConnectorImpl.findActiveContract("42", "42", "42"));
-
     }
 
     @Test
@@ -1261,11 +796,5 @@ class TokenConnectorImplTest {
         Pageable capturedPage = pageableArgumentCaptor.getValue();
         assertEquals(pageNumber, capturedPage.getPageNumber());
         verifyNoMoreInteractions(tokenRepository);
-    }
-
-    @Test
-    void countAllTokenFilterByStates(){
-        when(tokenRepository.count(any(), any())).thenReturn(1L);
-        Assertions.assertEquals(1L, tokenConnectorImpl.countAllTokenFilterByStates(List.of(RelationshipState.ACTIVE)));
     }
 }
