+++ core/src/main/java/it/pagopa/selfcare/mscore/core/strategy/input/CreateInstitutionStrategyInput.java
@@ -1,5 +1,6 @@
 package it.pagopa.selfcare.mscore.core.strategy.input;
 
+import it.pagopa.selfcare.commons.base.utils.InstitutionType;
 import it.pagopa.selfcare.mscore.core.util.InstitutionPaSubunitType;
 import it.pagopa.selfcare.mscore.model.institution.InstitutionGeographicTaxonomies;
 import lombok.Builder;
@@ -16,4 +17,5 @@ public class CreateInstitutionStrategyInput {
     private InstitutionPaSubunitType subunitType;
     private List<InstitutionGeographicTaxonomies> geographicTaxonomies;
     private String subunitCode;
+    private InstitutionType institutionType;
 }
