+++ core/src/test/java/it/pagopa/selfcare/mscore/core/OnboardingServiceImplTest.java
@@ -67,6 +67,9 @@ class OnboardingServiceImplTest {
     private ContractService contractService;
 
     @Mock
+    private ContractEventNotificationService contractEventNotificationService;
+
+    @Mock
     private UserService userService;
     @Mock
     private MailNotificationService emailService;
@@ -1683,12 +1686,60 @@ class OnboardingServiceImplTest {
         when(institutionConnector.findAndUpdate(any(), any(), any(), any())).thenReturn(institution);
 
         when(userService.retrieveUserFromUserRegistryByFiscalCode(userToOnboard.getTaxCode())).thenReturn(dummyUser());
-        when(userService.persistWorksContractToUserRegistry(any(), any(), any())).thenReturn(user);
+        when(userService.persistWorksContractToUserRegistry(userToOnboard.getTaxCode(), userToOnboard.getEmail(), institution.getId())).thenReturn(user);
 
         onboardingServiceImpl.persistOnboarding(institution.getId(), productId, userToOnboards, onboardingToPersist);
 
         verify(userService, times(1))
                 .retrieveUserFromUserRegistryByFiscalCode(userToOnboard.getTaxCode());
+
+        verify(userService, times(1))
+                .persistWorksContractToUserRegistry(userToOnboard.getTaxCode(), userToOnboard.getEmail(), institution.getId());
+
+        verify(onboardingDao, times(1))
+                .onboardOperator(any(), any(), any());
+        verifyNoMoreInteractions(userService);
+    }
+
+    /**
+     * Method under test: {@link OnboardingServiceImpl#onboardingUsers(OnboardingUsersRequest, String, String)}
+     */
+    @Test
+    void persistOnboarding_whenUserExistsOnRegistryButMailIsEmpty() {
+
+        String pricingPlan = "pricingPlan";
+        String productId = "productId";
+        Onboarding onboarding = dummyOnboarding();
+        onboarding.setStatus(UtilEnumList.VALID_RELATIONSHIP_STATES.get(0));
+
+        Onboarding onboardingToPersist = new Onboarding();
+        onboardingToPersist.setPricingPlan(pricingPlan);
+        onboardingToPersist.setProductId(productId);
+        onboardingToPersist.setBilling(new Billing());
+
+        Institution institution = new Institution();
+        institution.setId("institutionId");
+        institution.setOnboarding(List.of(onboarding));
+
+        UserToOnboard userToOnboard = createSimpleUserToOnboard();
+        userToOnboard.setEmail(null);
+        User user = new User();
+        user.setFiscalCode("fiscalCode");
+        user.setId("42");
+        final List<UserToOnboard> userToOnboards = List.of(userToOnboard);
+
+        when(institutionConnector.findById(institution.getId())).thenReturn(institution);
+        when(institutionConnector.findAndUpdate(any(), any(), any(), any())).thenReturn(institution);
+
+        when(userService.retrieveUserFromUserRegistryByFiscalCode(userToOnboard.getTaxCode())).thenReturn(dummyUser());
+
+        onboardingServiceImpl.persistOnboarding(institution.getId(), productId, userToOnboards, onboardingToPersist);
+
+        verify(userService, times(1))
+                .retrieveUserFromUserRegistryByFiscalCode(userToOnboard.getTaxCode());
+
+        verify(userService, times(0))
+                .persistWorksContractToUserRegistry(any(), any(), any());
 
         verify(onboardingDao, times(1))
                 .onboardOperator(any(), any(), any());
