+++ core/src/test/java/it/pagopa/selfcare/mscore/core/strategy/OnboardingInstitutionStrategyTest.java
@@ -46,6 +46,8 @@ class OnboardingInstitutionStrategyTest {
     @Mock
     private ContractService contractService;
     @Mock
+    private ContractEventNotificationService contractEventNotificationService;
+    @Mock
     private UserEventService userEventService;
     @Mock
     private UserService userService;
@@ -62,7 +64,7 @@ class OnboardingInstitutionStrategyTest {
     @BeforeEach
     void beforeAll() {
         strategyFactory = new OnboardingInstitutionStrategyFactory(onboardingDao,
-                contractService,userService, institutionService, coreConfig, mailNotificationService, fileStorageConnector, userEventService);
+                contractService, contractEventNotificationService, userService, institutionService, coreConfig, mailNotificationService, fileStorageConnector, userEventService);
     }
