+++ helm/values-prod.yaml
@@ -42,3 +42,4 @@ configEnvSpecific:
   APPLICATIONINSIGHTS_INSTRUMENTATION_LOGGING_LEVEL: "OFF"
   EXTERNAL_API_LOG_LEVEL: "INFO"
   CORE_USER_EVENT_SERVICE_TYPE: "send"
+  CORE_CONTRACT_EVENT_SERVICE_TYPE: "send"
