+++ helm/values-uat.yaml
@@ -35,3 +35,4 @@ configEnvSpecific:
   APPLICATIONINSIGHTS_INSTRUMENTATION_LOGGING_LEVEL: "OFF"
   EXTERNAL_API_LOG_LEVEL: "DEBUG"
   CORE_USER_EVENT_SERVICE_TYPE: "send"
+  CORE_CONTRACT_EVENT_SERVICE_TYPE: "send"
