+++ web/src/main/java/it/pagopa/selfcare/mscore/web/model/delegation/DelegationRequest.java
@@ -1,10 +1,12 @@
 package it.pagopa.selfcare.mscore.web.model.delegation;
 
+import it.pagopa.selfcare.mscore.constant.DelegationState;
 import it.pagopa.selfcare.mscore.constant.DelegationType;
 import lombok.Data;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+import java.time.OffsetDateTime;
 
 @Data
 public class DelegationRequest {
@@ -21,5 +23,8 @@ public class DelegationRequest {
     private String productId;
     @NotNull
     private DelegationType type;
+    private DelegationState status;
+    private OffsetDateTime createdAt;
+    private OffsetDateTime updatedAt;
 
 }
