+++ web/src/main/java/it/pagopa/selfcare/mscore/web/model/delegation/DelegationRequestFromTaxcode.java
@@ -1,10 +1,12 @@
 package it.pagopa.selfcare.mscore.web.model.delegation;
 
+import it.pagopa.selfcare.mscore.constant.DelegationState;
 import it.pagopa.selfcare.mscore.constant.DelegationType;
 import lombok.Data;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+import java.time.OffsetDateTime;
 
 @Data
 public class DelegationRequestFromTaxcode {
@@ -23,4 +25,7 @@ public class DelegationRequestFromTaxcode {
     private DelegationType type;
     private String fromSubunitCode;
     private String toSubunitCode;
+    private DelegationState status;
+    private OffsetDateTime createdAt;
+    private OffsetDateTime updatedAt;
 }
