+++ web/src/main/java/it/pagopa/selfcare/mscore/web/model/institution/InstitutionRequest.java
@@ -40,5 +40,6 @@ public class InstitutionRequest {
     private boolean imported;
     private OffsetDateTime createdAt;
     private OffsetDateTime updatedAt;
+    private boolean delegation;
 
 }
