+++ web/src/main/java/it/pagopa/selfcare/mscore/web/model/institution/InstitutionResponse.java
@@ -47,5 +47,6 @@ public class InstitutionResponse {
     private String aooParentCode;
     private OffsetDateTime createdAt;
     private OffsetDateTime updatedAt;
+    private boolean delegation;
 
 }
