+++ web/src/main/resources/swagger/swagger_en.properties
@@ -79,10 +79,12 @@ swagger.mscore.users.products=Retrieves products info and role which the user is
 swagger.mscore.users=Retrieves user given userId and optional ProductId
 swagger.mscore.users.delete.products=Delete logically the association institution and product
 swagger.mscore.api.users.updateUserStatus=Update user status with optional filter for institution, product, role and productRole
+swagger.mscore.api.users.getOnboardedUsers=Retrieve onboarded users according to identifiers in input
 swagger.mscore.institutions.delegations=Retrieve institution's delegations
 swagger.mscore.institutions.delegations.mode=Mode (full or normal) to retreieve institution's delegations
 swagger.mscore.institutions.brokers=Retrieve institution brokers
 swagger.mscore.users.userId=User's unique identifier
+swagger.mscore.users.userIds=Users unique identifiers
 swagger.mscore.api.users.updateUser=Service to send notification when user data get's updated
 swagger.mscore.institutions.api.getInstitutionUsers=Retrieve institution's users
 swagger.mscore.api.users.findAll=Retrieve all users according to optional params in input
